some improvements to selenium grid stuff
git-svn-id: https://trisano.csinitiative.net/svn/trisano/top-level/trunk@771 b21a84b9-e453-0410-92a3-c4f515f7ae45

+++ nedss-poc/rails/nedss/script/stopselgrid.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 export PATH=/usr/lib/firefox:$PATH
 
-killall Xvfb
+if [ "$HEADLESS" = "true" ]; then
+  echo "[Stopping X virtual frame buffer]"
+  killall Xvfb
+else
+  echo "[NOT stopping X virtual frame buffer]"
+fi
 
 # Uses MRI as JRuby doesn't spawn processes
+echo "[Stopping Selenium Grid Hub]"
 ruby -S rake -f $SELENIUM_GRID_HOME/Rakefile hub:stop & 
-ruby -S rake -f $SELENIUM_GRID_HOME/Rakefile rc:stop_all PORTS=5000-5009 & 
+sleep 2
+echo "[Stopping Selenium Grid RCs - port range: $PORTS]"
+ruby -S rake -f $SELENIUM_GRID_HOME/Rakefile rc:stop_all PORTS="$PORTS" &
